requirements.txt
Django==1.9.1
django-braces==1.8.1
Pillow==3.1.1
six==1.10.0
Unipath==1.1
wheel==0.26.0
